draft server v2

+++ internal/pkg/user/user_model.go
@@ -8,7 +8,7 @@ type User struct {
 	ID             int64          `db:"id"`
 	FirstName      string         `db:"first_name"`
 	SecondName     string         `db:"second_name"`
-	Age            sql.NullInt64  `db:"age"`
+	Age            sql.NullInt32  `db:"age"`
 	Biography      sql.NullString `db:"biography"`
 	City           sql.NullString `db:"city"`
 	Password       string         `db:"-"`
